tools/testing-cli/pyproject.toml
[project]
name = "kc-test"
version = "1.0.0"
description = "Testing CLI for Kong + Keycloak integration"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.1.7",
    "requests>=2.32.3",
    "python-jose>=3.3.0",
    "pydantic>=2.9.2",
    "rich>=13.9.2",
    "httpx>=0.27.2",
    "pyyaml>=6.0.1",
]

[project.optional-dependencies]
dev = ["pytest>=8.3.3", "pytest-cov>=5.0.0", "black>=24.0.0", "ruff>=0.6.0"]

[project.scripts]
kc-test = "cli:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ['py312']

[tool.ruff]
line-length = 100
target-version = "py312"
